+++ scripts/run_checks.py
@@ -3,22 +3,17 @@ from __future__ import annotations
 import sys
 from pathlib import Path
 
+from cognite.client import global_config
 from scripts.custom_checks.docstr_examples import format_docstring_examples
 from scripts.custom_checks.docstrings import format_docstrings
-from scripts.custom_checks.version import (
-    changelog_entry_date,
-    changelog_entry_version_matches,
-    pyproj_version_matches,
-    version_number_and_date_is_increasing,
-)
+from scripts.custom_checks.version import pyproj_version_matches
+
+global_config.silence_feature_preview_warnings = True
 
 
-def run_checks(files: list[Path]) -> list[str | None]:
+def run_checks(files: tuple[Path, ...]) -> list[str | None]:
     return [
         pyproj_version_matches(),
-        changelog_entry_version_matches(),
-        changelog_entry_date(),
-        version_number_and_date_is_increasing(),
         format_docstrings(files),
         format_docstring_examples(files),
     ]